+++ src/main/scala/Example.scala
@@ -1,4 +1,7 @@
-import lisa.KernelHelpers.{_, given}
+import utilities.KernelHelpers.{_, given}
+import utilities.tptp.ProblemGatherer.*
+import utilities.tptp.KernelParser.*
+import utilities.tptp.*
 import lisa.kernel.Printer.*
 import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
@@ -6,8 +9,6 @@ import lisa.kernel.proof.SCProofChecker
 import lisa.kernel.proof.SCProofChecker.*
 import lisa.kernel.proof.SequentCalculus.*
 import proven.tactics.SimplePropositionalSolver.solveSequent
-import tptp.KernelParser.*
-import tptp.ProblemGatherer.getPRPproblems
 
 /**
  * Discover some of the elements of LISA to get started.
