+++ src/main/scala/lisa/settheory/SetTheoryTGAxioms.scala
@@ -1,6 +1,6 @@
 package lisa.settheory
-import lisa.KernelHelpers.{_, given}
 import lisa.kernel.fol.FOL._
+import utilities.KernelHelpers.{given, _}
 
 /**
  * Axioms for the Tarski-Grothendieck theory (TG)
