+++ src/main/scala/proven/tactics/SimplePropositionalSolver.scala
@@ -1,11 +1,12 @@
 package proven.tactics
 
-import lisa.KernelHelpers.*
 import lisa.kernel.fol.FOL.*
 import lisa.kernel.proof.SCProof
 import lisa.kernel.proof.SequentCalculus.*
 
 import scala.collection.mutable.Set as mSet
+import utilities.KernelHelpers.*
+import lisa.kernel.proof.SCProof
 
 /**
  * A simple but complete solver for propositional logic. Will not terminate for large problems
