+++ src/test/scala/lisa/proven/SimpleProverTests.scala
@@ -1,6 +1,6 @@
 package proven
 
-import lisa.KernelHelpers.*
+import utilities.KernelHelpers.*
 import lisa.kernel.Printer
 import lisa.kernel.Printer.*
 import lisa.kernel.fol.FOL.*
@@ -8,9 +8,9 @@ import lisa.kernel.proof.RunningTheory
 import lisa.kernel.proof.RunningTheory.PredicateLogic
 import lisa.kernel.proof.SCProofChecker
 import org.scalatest.funsuite.AnyFunSuite
+import utilities.tptp.ProblemGatherer.getPRPproblems
+import utilities.tptp.KernelParser.*
 import proven.tactics.SimplePropositionalSolver as SPS
-import tptp.KernelParser.*
-import tptp.ProblemGatherer.getPRPproblems
 
 class SimpleProverTests extends AnyFunSuite {
